requirements.txt
aiogram==3.1.1
SQLAlchemy~=1.3.24
gino~=1.0.1
python-dotenv~=1.0.0
asyncpg~=0.29.0